+++ packages/rel8/junction/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./rel8-junction"
+export * from "./junction"
